+++ src/main/java/com/daylily/domain/auth/repository/UserRepository.java
@@ -9,5 +9,5 @@ import java.util.Optional;
 @Repository
 public interface UserRepository extends JpaRepository<User, Long> {
     // email은 비어있을 수 있으므로, 유니크 값인 githubId로 조회
-    Optional<User> findByGithubId(Long githubId);
+    Optional<User> findByGithubId(Integer githubId);
 }
